+++ 05_Essentials_Practice/01-starting-project/src/components/Header.jsx
@@ -3,7 +3,7 @@ import logo from '../assets/investment-calculator-logo.png';
 
 export default function Header() {
     return <header id="header">
-            <img src={logo} alt="ogo showing a money bag" />
+            <img src={logo} alt="logo showing a money bag" />
             <h1>Investment calculator</h1>
         </header>
     ;
