+++ 05_Essentials_Practice/01-starting-project/src/components/Results.jsx
@@ -1,0 +1,8 @@
+export default function Results({input}) {
+    console.log('Results        input', input);
+    return (
+        <div>
+            <h1>Results</h1>
+        </div>
+    )
+}
