+++ 05_Essentials_Practice/01-starting-project/src/components/UserInput.jsx
@@ -1,21 +1,4 @@
-import { useState } from 'react';
-
-export default function UserInput() {
-
-    const [userInput, setUserInput] = useState({
-        initialInvestment: 10000,
-        annualInvestment: 1200,
-        expectedReturn: 6,
-        duration: 10
-    });
-
-    function handleUserInputChange(userInputIdentifier, newValue) {
-        setUserInput({
-            ...userInput,
-            [userInputIdentifier]: newValue
-        });
-
-    }
+export default function UserInput({userInput, onChange}) {
 
     return (
         <section id="user-input">
@@ -25,7 +8,7 @@ export default function UserInput() {
                     <input type="number" 
                         required
                         value={userInput.initialInvestment}
-                        onChange={(event) => handleUserInputChange('initialInvestment', event.target.value)}
+                        onChange={(event) => onChange('initialInvestment', event.target.value)}
                     />
                 </p>
                 <p>
@@ -33,7 +16,7 @@ export default function UserInput() {
                     <input type="number" 
                         required
                         value={userInput.annualInvestment}
-                        onChange={(event) => handleUserInputChange('annualInvestment', event.target.value)}
+                        onChange={(event) => onChange('annualInvestment', event.target.value)}
                     />
                 </p>                
             </div>
@@ -43,7 +26,7 @@ export default function UserInput() {
                     <input type="number" 
                         required
                         value={userInput.expectedReturn}
-                        onChange={(event) => handleUserInputChange('expectedReturn', event.target.value)}
+                        onChange={(event) => onChange('expectedReturn', event.target.value)}
                     />
                 </p>
                 <p>
@@ -51,7 +34,7 @@ export default function UserInput() {
                     <input type="number" 
                         required
                         value={userInput.duration}
-                        onChange={(event) => handleUserInputChange('duration', event.target.value)}
+                        onChange={(event) => onChange('duration', event.target.value)}
                     />
                 </p>                
             </div>
